7_SEG_HEX
This is a sketch written for a hexadecimal number to be written to an LED. The sketch will count from zero to "F", going up by one per second. Once the program is done, the sketch will loop, and will repeat the process. Due to the limited number of pins on an Arduino, the program is limited to outputing the hexadecimal number in binary and WILL have to rely on a 4-Bit Binary to Hexadecimal multiplexer.